+++ src/Exceptions/DependencyNotMetException.php
@@ -1,0 +1,11 @@
+<?php
+namespace Michaels\Manager\Exceptions;
+
+/**
+ * DependencyNotMetException
+ * @package Michaels\Manager\Exceptions
+ */
+class DependencyNotMetException extends \Exception
+{
+
+}