+++ src/Traits/ChainsNestedItemsTrait.php
@@ -11,6 +11,8 @@ namespace Michaels\Manager\Traits;
  */
 trait ChainsNestedItemsTrait
 {
+    use DependsOnManagesItemsTrait;
+
     /**
      * Current level of nesting
      * @var bool|string
@@ -23,6 +25,7 @@ trait ChainsNestedItemsTrait
      */
     public function drop()
     {
+        $this->ensureCanManageItems();
         return $this->remove($this->currentLevel);
     }
 
@@ -35,6 +38,7 @@ trait ChainsNestedItemsTrait
      */
     public function __get($name)
     {
+        $this->ensureCanManageItems();
         $prefix = $this->buildPrefix();
 
         return $this->get($prefix . $name);
@@ -51,6 +55,7 @@ trait ChainsNestedItemsTrait
      */
     public function __call($name, $arguments)
     {
+        $this->ensureCanManageItems();
         $dot = ($this->currentLevel === false) ? '' : '.';
         $this->currentLevel .= $dot . $name;
 
@@ -67,6 +72,7 @@ trait ChainsNestedItemsTrait
      */
     public function __set($key, $value)
     {
+        $this->ensureCanManageItems();
         $prefix = $this->buildPrefix();
 
         return $this->add($prefix . $key, $value);
@@ -77,6 +83,8 @@ trait ChainsNestedItemsTrait
      */
     protected function buildPrefix()
     {
+        $this->ensureCanManageItems();
+
         if ($this->currentLevel === false) {
             return "";
         } else {