+++ tests/Stubs/ChainsNestedItemsTraitStub.php
@@ -1,6 +1,7 @@
 <?php
 namespace Michaels\Manager\Test\Stubs;
 
+use Michaels\Manager\Contracts\ManagesItemsInterface;
 use Michaels\Manager\Traits\ChainsNestedItemsTrait;
 use Michaels\Manager\Traits\ManagesItemsTrait;
 
@@ -8,7 +9,7 @@ use Michaels\Manager\Traits\ManagesItemsTrait;
  * Class ChainsNestedItemsTraitStub
  * @package Stubs
  */
-class ChainsNestedItemsTraitStub
+class ChainsNestedItemsTraitStub implements ManagesItemsInterface
 {
     use ManagesItemsTrait, ChainsNestedItemsTrait;